sshca: move the event handler to its own package
In order to move it we also needed to start passing in the context to
the callbacks.
